camera: move Camera_LoadCutsceneFrame to TRX
This moves both games' Camera_LoadCutsceneFrame functions to TRX as-is.
