+++ src/tr1/game/camera/common.h
@@ -3,5 +3,3 @@
 #include "global/types.h"
 
 #include <stdint.h>
-
-void Camera_UpdateCutscene(void);
