+++ src/tr2/game/camera.h
@@ -3,5 +3,3 @@
 #include "global/types.h"
 
 #include <libtrx/game/camera.h>
-
-void Camera_UpdateCutscene(void);
